+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2.0.0
+- Changed the code structure.
+- CLI support was added.
+- Added getting secondary location of any atsign.
+- Fixed keys folder issues.
+
+> Known issues: 
+> - Windows doesn't show any data while pasting digest challenge.
+> - Windows doesn't reflect any animations.
+
 ## 1.0.0
 
 A helper command